requirements.txt
aiogram==3.1.1
aiohttp==3.8.6
pydantic==2.3.0
python-dotenv==1.0.1
httpx==0.25.2
anyio==4.2.0
sniffio==1.3.1
annotated-types==0.7.0
aiofiles==23.1.0
magic-filter==1.0.12
certifi>=2023.7.22
